Allow specification of bind_address for port forwards
The previous behaviour was inconsistent:
For local and dynamic forwards the bind_address was set to localhost.
For remote forwards it was set to "", which means all interfaces / 0.0.0.0.
This could unintentionally expose the forwarded remote port to the outside,
in case GatewayPorts was enabled on the server side.

The bind_address is now fully customizable in the port forward dialog,
with "localhost" as default value for all new forwards.